refactor: collection_exchange.rs into smaller modules
Split the 648-line `profile/collection_exchange.rs` into a
`profile/collection_exchange/` module with single-responsibility files,
all well under the 500-line soft cap:

- `error.rs`    (82) — `CollectionExchangeError` enum and impls
- `types.rs`    (44) — public export/preview DTOs
- `matching.rs` (139) — local profile match index and descriptor classifier
- `export.rs`   (115) — TOML export + writer helper
- `import.rs`   (88) — preset parse + classify preview
- `mod.rs`      (244) — re-exports and test module

Public API is preserved; `profile::mod.rs` re-exports remain unchanged.

Part of #290
Closes #379
